z[3] & z[6] stay 0

+++ C_verzija/sha3_functions.c
@@ -82,10 +82,10 @@
                *(A[x][y]) ^= keccakf_rndc[n];
         }
     }
-    uint64_t rotate_right(uint64_t a, const uint64_t rotc)
+    int32_t rotate_right(int32_t a, const int32_t rotc)
     {
-        uint64_t rot = rotc % 64;
-        uint64_t tmp;
+        int32_t rot = rotc % 64;
+        int32_t tmp;
         while(rot)
         {   
             tmp = a & 0x8000000000000000;
